+++ apps/web/app/(ee)/api/partners/platforms/callback/route.ts
@@ -141,6 +141,10 @@ export async function GET(req: Request) {
         identifier,
       },
     },
+    select: {
+      id: true,
+      identifier: true,
+    },
   });
 
   if (!partnerPlatform || !partnerPlatform.identifier) {
