+++ apps/web/lib/api/partner-profile/upsert-partner-platform.ts
@@ -1,5 +1,4 @@
 import { prisma } from "@/lib/prisma";
-import { MAX_PLATFORMS_PER_TYPE } from "@/lib/social-utils";
 import { PartnerPlatform, Prisma } from "@prisma/client";
 
 type UpsertPartnerPlatformParams = {
@@ -32,35 +31,4 @@ export async function upsertPartnerPlatform({
       ...data,
     },
   });
-}
-
-export async function assertPartnerPlatformLimit({
-  partnerId,
-  type,
-  identifier,
-}: Pick<PartnerPlatform, "partnerId" | "type" | "identifier">) {
-  const existingPlatforms = await prisma.partnerPlatform.findMany({
-    where: {
-      partnerId,
-      type,
-    },
-    select: {
-      identifier: true,
-    },
-  });
-
-  // Re-verifying / updating an existing handle is always allowed
-  const alreadyExists = existingPlatforms.some(
-    (p) => p.identifier === identifier,
-  );
-
-  if (alreadyExists) {
-    return;
-  }
-
-  if (existingPlatforms.length >= MAX_PLATFORMS_PER_TYPE) {
-    throw new Error(
-      `You can add up to ${MAX_PLATFORMS_PER_TYPE} ${type} handles.`,
-    );
-  }
 }
